client/requirements.txt
pygame
pystray
python-socketio[client]
requests
Pillow
python-dotenv
toml
